# apiEvaluacion/api-docs.json
{"swagger":"2.0","info":{"description":"Servicios Rest para la calificación de evaluaciones","version":"API v1","title":"RestApi Evaluación","termsOfService":"Terminos del servicio","contact":{"name":"Edwin Quispe","url":"https://www.linkedin.com/in/edwinqramos/","email":"edwinqramos@gmail.com"},"license":{"name":"Licencia del API","url":"API licencia URL"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"api-evaluacion","description":"Api Evaluacion"}],"paths":{"/api/v1/evaluaciones/listadoRangoFechas/{fechaInicio}/{fechaFin}":{"get":{"tags":["api-evaluacion"],"summary":"Listado de evaluaciones por Rango de fechas","operationId":"listadoRangoFechasUsingGET","produces":["application/json"],"parameters":[{"name":"fechaFin","in":"path","description":"fechaFin","required":true,"type":"string","x-example":"19-09-2020"},{"name":"fechaInicio","in":"path","description":"fechaInicio","required":true,"type":"string","x-example":"19-09-2020"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Evaluacion"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"},"500":{"description":"Internal Error Server"}}}},"/api/v1/evaluaciones/modificaciones":{"put":{"tags":["api-evaluacion"],"summary":"Modifica una evaluación","operationId":"modificarUsingPUT","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"evaluacion","description":"evaluacion","required":true,"schema":{"$ref":"#/definitions/Evaluacion"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RespuestaApi"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"},"500":{"description":"Internal Error Server","schema":{"$ref":"#/definitions/RespuestaApi"}}}}},"/api/v1/evaluaciones/registros":{"post":{"tags":["api-evaluacion"],"summary":"Registra una nueva evaluación","operationId":"registrarUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"evaluacion","description":"evaluacion","required":true,"schema":{"$ref":"#/definitions/Evaluacion"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RespuestaApi"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"},"500":{"description":"Internal Error Server","schema":{"$ref":"#/definitions/RespuestaApi"}}}}}},"definitions":{"Evaluacion":{"type":"object","properties":{"calificacion":{"type":"integer","format":"int32"},"email":{"type":"string"},"fechaCreacion":{"type":"string","format":"date-time"},"idEvaluacion":{"type":"integer","format":"int64"},"nombres":{"type":"string"}},"title":"Evaluacion"},"RespuestaApi":{"type":"object","properties":{"body":{"type":"object"},"status":{"type":"string"}},"title":"RespuestaApi"}}}